fix : conflict  중 오류해결
