+++ src/assets/scss/components/_index.scss
@@ -4,4 +4,4 @@
 @import "./modal/baseModal";
 @import "./modal/createRoomModal";
 @import "./modal/entryRoomModal";
-
+@import "./createRoom/editorModeButton";
