+++ src/components/index.ts
@@ -8,4 +8,4 @@ import CodeModeSelect from "./createRoom/CodeModeSelect";
 import PdfModeSelect from "./createRoom/PdfModeSelect";
 import EditorModeButton from "./createRoom/EditorModeButton";
 
-export {Navbar,Button ,BlankScreenModeSelect ,CodeModeSelect,PdfModeSelect,EditorModeButton};
+export {Navbar,Button ,BlankScreenModeSelect ,CodeModeSelect,PdfModeSelect,EditorModeButton, SelectButton,BaseModal ,EntryModal};
